Refactor and add equipment to units in the demo state
