+++ src/utils/mod.rs
@@ -1,2 +1,3 @@
+pub mod camera;
 pub mod movement;
 pub mod sprites;
